+++ src/vec.rs
@@ -117,10 +117,10 @@ impl Vec3 {
 
 impl Color {
     /// Get color of pixel by taking the average over the number of `samples`.
-    pub fn format_color(self, samples: u64) -> String {
-        let ir = (256.0 * (self[0] / (samples as f64)).sqrt().clamp(0.0, 0.999)) as u64;
-        let ig = (256.0 * (self[1] / (samples as f64)).sqrt().clamp(0.0, 0.999)) as u64;
-        let ib = (256.0 * (self[2] / (samples as f64)).sqrt().clamp(0.0, 0.999)) as u64;
+    pub fn format_color(self, samples: u32) -> String {
+        let ir = (256.0 * (self[0] / f64::from(samples)).sqrt().clamp(0.0, 0.999)) as u32;
+        let ig = (256.0 * (self[1] / f64::from(samples)).sqrt().clamp(0.0, 0.999)) as u32;
+        let ib = (256.0 * (self[2] / f64::from(samples)).sqrt().clamp(0.0, 0.999)) as u32;
 
         format!("{} {} {}", ir, ig, ib)
     }
